+++ internal/provider/kubernetes_cluster_resource_acc_test.go
@@ -83,7 +83,6 @@ resource "prodata_kubernetes_cluster" "test" {
   kubernetes_version = data.prodata_kubernetes_versions.v.latest_version
   network_id         = prodata_local_network.k8s.id
   pod_cidr           = "10.244.0.0/16"
-  node_subnet        = 24
   node_ip_range      = "10.30.0.10-10.30.0.20"
   master_flavor_id   = data.prodata_kubernetes_flavors.standard.flavors[0].id